pyproject.toml
[project]
name = "ai-employee-vault"
version = "0.2.0"
description = "Silver Tier AI Employee - Multi-Channel Integration"
requires-python = ">=3.13"
dependencies = [
    "watchdog>=4.0.0",
    "google-auth>=2.23.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.1.1",
    "google-api-python-client>=2.100.0",
    "playwright>=1.40.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
